Add files via upload
Shift registers are digital circuits that are commonly used in a variety of electronic applications to store and transfer data. RTL stands for Register Transfer Level, which is a design abstraction used in digital circuit design. RTL design describes the behavior of a circuit in terms of the flow of data between registers.

This repository contains RTL and TB codes for shift registers would be a collection of digital circuit design files that implement shift registers using the RTL design abstraction and test the functionality of the design using testbench modules. This repository could be used by digital circuit designers as a reference or as a starting point for their own shift register designs.